requirements.txt
azure-functions
azure-cosmos
